+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Change Log
 - Fix an issue on CotEditor 4.4.0 that `contents of selection` of a document object returned its entire contents.
 - Fix an issue that the last of multiple insertion points locates at the end of the document was not drawn.
 - Fix an issue that the insertion point immediately exited automatically closed quote marks when the smart quote feature is enabled.
+- Fix an issue that the initial window height could differ from the user-specified setting.